+ proper dungeon rendering made the dungeon render in a more dynammic way made the player start in the middle of the dungeon

+++ Dungeon-Shambles/Dungeon-Shambles/Dungeon Generation/Room.cs
@@ -60,7 +60,7 @@ namespace DungeonShambles
 			}
 		}
 
-		public void renderRoom(int offsetX, int offsetY) {
+		public void renderRoom(float offsetX, float offsetY) {
 			for (int i = 0; i < roomWidth; i++) {
 				for (int j = 0; j < roomHeight; j++) {
 					tiles [i,j].renderTile (Globals.TextureSize, i * Globals.TextureSize * 2 + offsetX, j * Globals.TextureSize * 2 + offsetY);
